% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco.R
\name{coco}
\alias{coco}
\title{coco
#This is a function to test.}
\usage{
coco(a, b)
}
\arguments{
\item{FLOAT}{OR INTEGER > 0}
}
\value{
FLOAT or INTEGER >0
}
\description{
coco
#This is a function to test.
}
\examples{
coco(1,2)

}
