% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWN.R
\name{CWN}
\alias{CWN}
\title{Delete NAs from the dataframe}
\usage{
CWN(df)
}
\arguments{
\item{df}{any dataframes}
}
\value{
Dataframe that without NAs
}
\description{
Delete NAs from the dataframe
}
\examples{
CWN(mtcars)
}
