% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HealthData.R
\docType{data}
\name{HealthData}
\alias{HealthData}
\title{WORLD BANK DATA}
\format{
An object of class \code{data.frame} with 259 rows and 69 columns.
}
\usage{
HealthData
}
\description{
WORLD BANK DATA
}
\keyword{datasets}
