% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly-beer-production-in-austr.R
\docType{data}
\name{Monthly_beer_production_austr}
\alias{Monthly_beer_production_austr}
\title{monthly-beer-production-in-austr}
\format{
477 x 2 dataframe ...

\describe{
\item{index}{Time}
\item{column1}{Monthly beer production}
}
}
\usage{
Monthly_beer_production_austr
}
\description{
This is time series data about the montly beer production in Australia.
}
\keyword{datasets}
