% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plateprice.R
\docType{data}
\name{Plateprice}
\alias{Plateprice}
\title{Shanghai License Plate Deal Price from 2000 to 2017}
\format{
190 x 6 dataframe ...

\describe{
\item{index}{ID}
\item{column1}{Date}
\item{column2}{avg_deal_price （$)}
\item{column3}{lowest_deal_price ($)}
\item{column4}{num_bidder}
\item{column5}{num_plates}
}
}
\source{
https://www.kaggle.com/bazingasu/shanghai-license-plate-bidding-price-prediction
}
\usage{
Plateprice
}
\description{
Shanghai License Plate Deal Price from 2000 to 2017, all of them have been converted to U.S dollar price
}
\keyword{datasets}
