% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_transfrom_test.R
\name{pre_transform_test}
\alias{pre_transform_test}
\title{To test whether the variables in the liner regression model needs a transform(bcPower or yjPower) to generate a better model.}
\usage{
pre_transform_test(n)
}
\arguments{
\item{n}{it will be any varibales in a liner regression model}
}
\value{
Estimated transformation parameter
}
\description{
To test whether the variables in the liner regression model needs a transform(bcPower or yjPower) to generate a better model.
}
\examples{
pre_transfrom_test(mtcars)
}
