% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recursive_backtest.R
\name{Recursive_backtest}
\alias{Recursive_backtest}
\title{Recursive_backtest}
\usage{
Recursive_backtest(Df_ts, x, y, z)
}
\arguments{
\item{Df_ts}{Any time series data.}

\item{x}{Numbers of forecast, x+y is equal to the total number of the time series.}

\item{y}{The beginning of the selected estimation data.}

\item{z}{How much steps you want to forecast (you need to take care that z+y+i cannot excess the total number of the time series )}
}
\value{
The forecast result of the recursive_backtest.
}
\description{
Recursive_backtest
}
\examples{
Recursive_backtest(any_time_series_data,3,4,1)
}
