% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top_Car_Sales.R
\docType{data}
\name{Top_Car_Sales}
\alias{Top_Car_Sales}
\title{Top Car sales from 2017 - 2000}
\format{

}
\source{
https://www.kaggle.com/italomarcelo/top-cars-sales-20182020-fictitious
}
\usage{
Top_Car_Sales
}
\description{
Top Car sales from 2017 - 2000
}
\keyword{datasets}
